# apps/api/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Yougan API",
    "version": "0.1.0",
    "description": "有感 Yougan 中间层：账号鉴权、作品管理、OSS 上传、Agent 流转发。"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "PlatformIntegration": {
        "type": "object",
        "nullable": true,
        "properties": {
          "id": {
            "type": "string"
          },
          "platform": {
            "type": "string"
          },
          "accountName": {
            "type": "string",
            "nullable": true
          },
          "accountId": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string"
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tokenExpiresAt": {
            "type": "string",
            "nullable": true
          },
          "connectedAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "platform",
          "accountName",
          "accountId",
          "status",
          "scopes",
          "tokenExpiresAt",
          "connectedAt",
          "updatedAt"
        ]
      },
      "PlatformCatalogItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "oauthConfigured": {
            "type": "boolean"
          },
          "connected": {
            "type": "boolean"
          },
          "integration": {
            "$ref": "#/components/schemas/PlatformIntegration"
          }
        },
        "required": [
          "id",
          "label",
          "description",
          "oauthConfigured",
          "connected",
          "integration"
        ]
      },
...
      "WorkProfile": {
        "type": "object",
        "properties": {
          "platform": {
            "type": "string",
            "nullable": true
          },
          "content_topic": {
            "type": "string",
            "nullable": true
          },
          "content_type": {
            "type": "string",
            "nullable": true
          },
          "content_points": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "style": {
            "type": "string",
            "nullable": true
          },
          "tone": {
            "type": "string",
            "nullable": true
          },
          "persona": {
            "type": "string",
            "nullable": true
          },
          "style_constraints": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "audience": {
            "type": "string",
            "nullable": true
          },
          "goals": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "references": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": {
                "nullable": true
              }
            }
          }
        }
      },
      "WorkOutline": {
        "type": "object",
        "properties": {
          "pending_changes": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "created_at": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "description",
                "created_at"
              ]
            }
          },
          "executed_changes": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "executed_at": {
                  "type": "string"
                },
                "batch_summary": {
                  "type": "string",
                  "nullable": true
                }
              },
              "required": [
                "id",
                "description",
                "executed_at"
              ]
            }
          },
          "last_execution_summary": {
            "type": "string",
            "nullable": true
          },
          "outline_summary": {
            "type": "string",
            "nullable": true
          },
          "outline_ready": {
            "type": "boolean"
          }
        },
        "required": [
          "pending_changes",
          "executed_changes"
        ]
      },
      "WorkInspiration": {
        "type": "object",
        "properties": {
          "confirmed_requirements": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "confirmed_at": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "description",
                "confirmed_at"
              ]
            }
          },
          "summary": {
            "type": "string",
            "nullable": true
          },
          "inspiration_ready": {
            "type": "boolean"
          },
          "summarized_at": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "confirmed_requirements"
        ]
      },
      "GeneratedContent": {
        "type": "object",
        "nullable": true,
        "properties": {
          "platform": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "body": {
            "type": "string"
          },
          "hashtags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "hook": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "images": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": {
                "nullable": true
              }
            }
          },
          "publish_ready": {
            "type": "boolean"
          }
        },
        "required": [
          "platform",
          "body"
        ]
      },
      "Work": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "groupId": {
            "type": "string",
            "nullable": true
          },
          "mode": {
            "type": "string",
            "enum": [
              "inspiration",
              "outline",
              "creation"
            ]
          },
          "threadId": {
            "type": "string",
            "nullable": true
          },
          "profile": {
            "$ref": "#/components/schemas/WorkProfile"
          },
          "outline": {
            "$ref": "#/components/schemas/WorkOutline"
          },
          "inspiration": {
            "$ref": "#/components/schemas/WorkInspiration"
          },
          "creation": {
            "$ref": "#/components/schemas/GeneratedContent"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "groupId",
          "mode",
          "threadId",
          "profile",
          "outline",
          "inspiration",
          "creation",
          "createdAt",
          "updatedAt"
        ]
      },
      "SyncWorkState": {
        "type": "object",
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "inspiration",
              "outline",
              "creation"
            ]
          },
          "threadId": {
            "type": "string",
            "nullable": true
          },
          "groupId": {
            "type": "string",
            "nullable": true
          },
          "profile": {
            "$ref": "#/components/schemas/WorkProfile"
          },
          "outline": {
            "$ref": "#/components/schemas/WorkOutline"
          },
          "inspiration": {
            "$ref": "#/components/schemas/WorkInspiration"
          },
          "creation": {
            "$ref": "#/components/schemas/GeneratedContent"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "AgentContext": {
        "type": "object",
        "properties": {
          "workId": {
            "type": "string"
          },
          "mode": {
            "type": "string",
            "enum": [
              "inspiration",
              "outline",
              "creation"
            ]
          },
          "profile": {
            "$ref": "#/components/schemas/WorkProfile"
          },
          "outline": {
            "$ref": "#/components/schemas/WorkOutline"
          },
          "inspiration": {
            "$ref": "#/components/schemas/WorkInspiration"
          },
          "creation": {
            "$ref": "#/components/schemas/GeneratedContent"
          }
        },
        "required": [
          "workId",
          "mode",
          "profile",
          "outline",
          "inspiration",
          "creation"
        ]
      },
      "InspirationRecommendation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "suggestion": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "suggestion"
        ]
      },
      "WorkInspirationRecommendations": {
        "type": "object",
        "properties": {
          "workId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "recommendations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InspirationRecommendation"
            },
            "minItems": 1,
            "maxItems": 5
          }
        },
        "required": [
          "workId",
          "title",
          "recommendations"
        ]
      },
...
      "SubscriptionSummary": {
        "type": "object",
        "properties": {
          "planId": {
            "type": "string"
          },
          "planName": {
            "type": "string"
          },
          "planDescription": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "billingCycle": {
            "type": "string",
            "nullable": true,
            "enum": [
              "monthly",
              "yearly"
            ]
          },
          "currentPeriodStart": {
            "type": "string"
          },
          "currentPeriodEnd": {
            "type": "string",
            "nullable": true
          },
          "aiUsageThisPeriod": {
            "type": "integer"
          },
          "aiQuotaThisPeriod": {
            "type": "integer"
          },
          "cancelAtPeriodEnd": {
            "type": "boolean"
          },
          "features": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "planId",
          "planName",
          "planDescription",
          "status",
          "billingCycle",
          "currentPeriodStart",
          "currentPeriodEnd",
          "aiUsageThisPeriod",
          "aiQuotaThisPeriod",
          "cancelAtPeriodEnd",
          "features"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/integrations": {
      "get": {
        "tags": [
          "Integrations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Platform integration catalog",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "platforms": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PlatformCatalogItem"
                      }
                    }
                  },
                  "required": [
                    "platforms"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/integrations/{platform}/authorize": {
      "post": {
        "tags": [
          "Integrations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "platform",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "OAuth authorization URL",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "authorizationUrl": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "authorizationUrl"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/integrations/{platform}": {
      "delete": {
        "tags": [
          "Integrations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "platform",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Disconnected"
          }
        }
      }
    },
...
    "/api/works": {
      "get": {
        "tags": [
          "Works"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List works",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "works": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Work"
                      }
                    }
                  },
                  "required": [
                    "works"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Works"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "groupId": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "work": {
                      "$ref": "#/components/schemas/Work"
                    }
                  },
                  "required": [
                    "work"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/works/{workId}": {
      "get": {
        "tags": [
          "Works"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "workId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Work detail",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "work": {
                      "$ref": "#/components/schemas/Work"
                    }
                  },
                  "required": [
                    "work"
                  ]
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Works"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "workId",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncWorkState"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "work": {
                      "$ref": "#/components/schemas/Work"
                    }
                  },
                  "required": [
                    "work"
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Works"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "workId",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          }
        }
      }
    },
    "/api/works/{workId}/agent-context": {
      "get": {
        "tags": [
          "Works"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "workId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Agent context for a work",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "context": {
                      "$ref": "#/components/schemas/AgentContext"
                    }
                  },
                  "required": [
                    "context"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/works/{workId}/inspiration-recommendations": {
      "post": {
        "tags": [
          "Works"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "workId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Recommended inspirations for a work",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkInspirationRecommendations"
                }
              }
            }
          },
          "404": {
            "description": "Work not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "Recommendation agent failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "DeepSeek not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/billing/subscription": {
      "get": {
        "tags": [
          "Billing"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current subscription",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "subscription": {
                      "$ref": "#/components/schemas/SubscriptionSummary"
                    }
                  },
                  "required": [
                    "subscription"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/billing/subscription/cancel": {
      "post": {
        "tags": [
          "Billing"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Subscription set to cancel at period end",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "subscription": {
                      "$ref": "#/components/schemas/SubscriptionSummary"
                    }
                  },
                  "required": [
                    "subscription"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/billing/subscription/resume": {
      "post": {
        "tags": [
          "Billing"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Subscription renewal resumed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "subscription": {
                      "$ref": "#/components/schemas/SubscriptionSummary"
                    }
                  },
                  "required": [
                    "subscription"
                  ]
                }
              }
            }
          }
        }
      }
    }
  }
}